moved firebase out of component folder
