+++ src/Components/FirebaseInteraction.js
@@ -1,4 +1,4 @@
-import firebase from "./firebase";
+import firebase from "../firebase";
 
 const FirebaseInteraction  = ({movieDetails}) => {
     const dbRef = firebase.database().ref();
